% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecan_version.R
\name{pecan_version}
\alias{pecan_version}
\title{Report installed PEcAn version}
\usage{
pecan_version(version = max(PEcAn.all::pecan_releases$version), exact = FALSE)
}
\arguments{
\item{version}{PEcAn release number to use for expected package versions}

\item{exact}{Show only tags that exactly match \code{version},
or all tags that have it as a substring?}
}
\value{
data frame with columns for package name, expected version(s),
and installed version
}
\description{
Reports the currently installed version of each PEcAn package, as well the
versions expected by the specified release(s) of the PEcAn system.
}
\details{
When \code{exact} = FALSE (the default), \code{pecan_version} will show all releases
whose names contain \code{version} as a substring.
This is mostly so that e.g. "1.4.2" finds the release tag "v1.4.2",
but since all matches are returned it can also be handy for seeing all
patches sharing a minor version number (see examples).
}
\examples{
pecan_version()
pecan_version(c("1.4.0", "1.5.0", "1.6.0"))

# multiple matches are possible when exact = FALSE
pecan_version("v1.3", exact = TRUE)
pecan_version("v1.3", exact = FALSE)

}
