% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pecan_version_history}
\alias{pecan_version_history}
\title{Versions of all PEcAn packages in each release of PEcAn}
\format{
Data frame with at least 33 columns (plus one more every release!)
\describe{
\item{package}{Package name, as character}
\item{v1.0, ..., v1.7.2, ....}{Release numbers,
each with the Git tag as the column name and numeric package versions
as the row contents.
NA means \code{package} did not exist in this version of PEcAn.}
}
}
\source{
Extracted from Git history of the
\href{https://github.com/PecanProject/pecan}{PEcAn development repository}
}
\usage{
pecan_version_history
}
\description{
Versions of all PEcAn packages in each release of PEcAn
}
\seealso{
\code{pecan_version}, which queries this table and also reports the
versions currently installed
}
\keyword{datasets}
